Restructure and auto-discover/collect tasks
[no ci integration]
